Change way pieces and level objects are initialized
Fix issue where setting the top or left value of a piece to 0 would cause it to be instantiated incorrectly since it assumed 0 was a null value. Change create piece method to pass in left offset instead of computing it in the piece object. Level object now gives presidence to a structure array passed in to determine width and height. Refactor some unit tests.